+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 
-Lua error on login fixed. Maybe.
+* LEFT and RIGHT added as options for the tooltip alignment.
+* New UI Element added under ESC -> Interface Options -> Token Reward Helper to ease the alignment setting.
+* `/trh align` with no valid alignment as parameter now opens the new alignment panel.
+* `/trh align <TOP|BOTTOM|LEFT|RIGHT>` is working as before and setting the alignment immediatly without opening an UI element.
+* Default alignment for new installations changed to BOTTOM instead of TOP.